requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
reportlab==4.0.7
pypdf2==3.0.1
anthropic==0.7.8
python-multipart==0.0.6
email-validator==2.1.0
phonenumbers==8.13.26
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
